requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-choices==1.7.2
django-crispy-forms==1.9.2
django-crum==0.7.9
django-environ==0.4.5
django-extra-views==0.14.0
django-health-check==3.16.5
gunicorn==20.0.4
mysqlclient==2.0.3
Pillow==8.2.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
